app/requirements.txt

flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
bcrypt==4.1.3
gunicorn==22.0.0
